requirements.txt
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
google-api-core==2.11.0
google-api-python-client==2.69.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
gspread==5.7.2
httplib2==0.21.0
idna==3.4
oauth2client==4.1.3
oauthlib==3.2.2
pixelblaze-client==1.0.2
playsound==1.3.0
protobuf==4.21.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.13
websocket-client==1.4.2
